[server] Fix mysql connection timeouts
This happened when the connection wasn't used for a few hours (the default mysql
connection timeout), and the next time anyone uses it, it crashes the app. This
needs to be fixed in 2 ways: automatically regain the connection, and also
a connection failure shouldn't crash the app. This commit fixes the first by
using a connection pool which automatically reconnects.